back-django-rest/requirements.txt
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.2
Django==2.2.26
django-allauth==0.44.0
django-cors-headers==3.6.0
django-environ==0.8.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
google-api-core==1.31.5
google-auth==1.35.0
google-cloud-core==2.2.2
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
idna==3.3
mysqlclient==2.1.0
oauthlib==3.1.1
packaging==21.3
Pillow==9.0.0
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyftpdlib==1.5.6
PyJWT==2.3.0
pyparsing==3.0.7
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
